pyproject.toml
[build-system]
requires = ["hatchling","hatch-vcs","wheel"]
build-backend = "hatchling.build"

[project]
name = "rd-tomo"
description = "Toolbox for tomographic SAR processing from RD350 drone data"
authors = [
    {name = "Ansgar Jörgenfelt", email = "ansgar.jorgenfelt@slu.se"}
]
keywords = ["SAR", "tomographic", "tomography", "tomogram", "Radaz", "RD350", "radar", "microwaves", "processing"]
readme = "README.md"
license = "BSD-3-Clause"
license-files = ["LICENSE"]
requires-python = ">=3.12"
dependencies = [
    "numpy (>=2.3.3,<3.0.0)",
    "pandas (>=2.3.3,<3.0.0)",
    "rasterio (>=1.4.3,<2.0.0)",
    "scipy (>=1.16.2,<2.0.0)",
    "scikit-learn (>=1.7.1,<2.0.0)",
    "matplotlib (>=3.10.6,<4.0.0)",
    "pyproj (>=3.7.2,<4.0.0)",
    "geopandas (>=1.1.1,<2.0.0)",
    "tqdm (>=4.67.1,<5.0.0)",
    "pytz (>=2025.2,<2026.0)",
    "scikit-image (>=0.25.2,<0.26.0)",
    "sympy (>=1.14.0,<2.0.0)",
    "click (>=8.2.1,<9.0.0)",
    "threadpoolctl (>=3.6.0,<4.0.0)",
    "rich (>=14.1.0,<15.0.0)",
    "hatch (>=1.5.1, <2.0.0)",
    "PyQt5 (>=5.15.11, <6.0.0)",
    "laspy (>=2.6.1, <3.0.0)",
    "pyyaml (>=6.0.3, <7.0.0)",
]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Programming Language :: Python"
]
dynamic=[
    "version",
]

[project.optional-dependencies]
dev = ["fawltydeps", "vermin", "pyubx2"]

[project.scripts]
rdtomo = "rdtomo.toolbox.rdtomo:rdtomo"

[project.urls]
Homepage = "https://github.com/ansgarj/rd-tomo.git"
Repository = "https://github.com/ansgarj/rd-tomo.git"
Changelog = "https://github.com/ansgarj/rd-tomo/blob/main/CHANGELOG.md"

[tool.hatch.version]
source = "vcs"
fallback-version = "0.0.0"

[tool.hatch.version.raw-options]
version_scheme = "no-guess-dev"

[tool.hatch.build.hooks.vcs]
version-file = "rdtomo/version.py"